refactor: flatten CLI prompts location
