+++ src/cli.ts
@@ -5,7 +5,7 @@ import { cac } from 'cac';
 
 import { logError, logger } from './logger.js';
 import { run, listVersions } from './index.js';
-import { promptForDirectory, promptForFigmaUrl, promptForVersions } from './cli/prompts.js';
+import { promptForDirectory, promptForFigmaUrl, promptForVersions } from './cli-prompts.js';
 
 interface CliFlags {
   list?: boolean;
